requirements.txt
apns2==0.7.1
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.2
cryptography==3.0
decorator==4.4.2
Django==3.0.8
django-appconf==1.0.4
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap4==2.2.0
django-classy-tags==2.0.0
django-cookie-law==2.0.3
django-encrypted-id==0.3.2
django-fontawesome-5==1.0.18
django-google-maps==0.12.2
django-grappelli==2.14.2
django-image-cropping==1.4.0
django-localflavor==3.0.1
django-location-field==2.1.0
django-ordered-model==3.4.1
django-recaptcha==2.0.6
django-richtextfield==1.6
django-walletpass==1.2
djangorestframework==3.11.1
easy-thumbnails==2.7
geographiclib==1.50
geopy==2.0.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
importlib-metadata==1.7.0
mysqlclient==2.0.1
Pillow==7.2.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.8
PyJWT==1.7.1
pypng==0.0.20
PyQRCode==1.2.1
pytest-runner==5.2
python-stdnum==1.13
python-telegram-bot==12.8
pytz==2020.1
reportlab==3.5.44
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tornado==6.0.4
zipp==3.1.0
